actually pass switches to test
