+++ mix.exs
@@ -3,7 +3,7 @@ defmodule DistributedTest.Mixfile do
 
   def project do
     [app: :distributed_test,
-     version: "0.1.0",
+     version: "0.2.0",
      elixir: "~> 1.4",
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
